Rename factory to store.
